requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
black==23.7.0
    # via -r requirements.in
cfgv==3.4.0
    # via pre-commit
click==8.1.7
    # via black
coverage[toml]==7.3.1
    # via pytest-cov
distlib==0.3.7
    # via virtualenv
filelock==3.12.4
    # via virtualenv
identify==2.5.29
    # via pre-commit
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via -r requirements.in
joblib==1.3.2
    # via scikit-learn
mypy==1.5.1
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nodeenv==1.8.0
    # via pre-commit
numpy==1.25.2
    # via
    #   -r requirements.in
    #   pandas
    #   scikit-learn
    #   scipy
packaging==23.1
    # via
    #   black
    #   pytest
pandas==2.0.3
    # via -r requirements.in
pathspec==0.11.2
    # via black
platformdirs==3.10.0
    # via
    #   black
    #   virtualenv
pluggy==1.3.0
    # via pytest
pre-commit==3.3.3
    # via -r requirements.in
pytest==7.4.0
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-mock
pytest-cov==4.1.0
    # via -r requirements.in
pytest-mock==3.11.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via pre-commit
ruff==0.0.284
    # via -r requirements.in
scikit-learn==1.3.0
    # via -r requirements.in
scipy==1.11.1
    # via
    #   -r requirements.in
    #   scikit-learn
six==1.16.0
    # via python-dateutil
threadpoolctl==3.2.0
    # via scikit-learn
typing-extensions==4.7.1
    # via mypy
tzdata==2023.3
    # via pandas
virtualenv==20.24.5
    # via pre-commit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
